requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
filelock==3.14.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipykernel==6.29.4
ipython==8.25.0
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.2.2
matplotlib-inline==0.1.7
multidict==6.0.5
nest-asyncio==1.6.0
openai==1.30.3
packaging==24.0
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
Protego==0.3.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pyzmq==26.0.3
queuelib==1.7.0
requests==2.31.0
requests-file==2.1.0
runstats==2.0.0
Scrapy==2.11.2
scrapy-zyte-api==0.18.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
tenacity==8.3.0
tldextract==5.1.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
Twisted==24.3.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
w3lib==2.1.2
wcwidth==0.2.13
yarl==1.9.4
zope.interface==6.4.post2
zyte-api==0.6.0
